fix: allow 'build'/'unit-tests' make targets to work on non-linux machines

do not default to GOOS=linux in Makefile

This reduces friction for running make targets on non-linux machines.
For example, the `validate-csv` target would fail prior to this commit
with the following error message:

```
CGO_ENABLED=0 GOOS=linux \
                go build -ldflags "-s -w -X github.com/NVIDIA/gpu-operator/internal/info.gitCommit=9aa8cc642 -X github.com/NVIDIA/gpu-operator/internal/info.version=v26.7.0"  github.com/NVIDIA/gpu-operator/cmd/gpuop-cfg
./gpuop-cfg validate csv --input=./bundle/manifests/gpu-operator-certified.clusterserviceversion.yaml
./gpuop-cfg: ./gpuop-cfg: cannot execute binary file
make: *** [validate-csv] Error 126
```

This change should not affect our GHA jobs as they all run on
linux runners, and therefore, GOOS should be set to linux there.
