+++ cmd/nvidia-validator/main.go
@@ -31,7 +31,6 @@ import (
 	"github.com/NVIDIA/go-nvlib/pkg/nvmdev"
 	"github.com/NVIDIA/go-nvlib/pkg/nvpci"
 	devchar "github.com/NVIDIA/nvidia-container-toolkit/cmd/nvidia-ctk/system/create-dev-char-symlinks"
-	"github.com/cyphar/filepath-securejoin/pathrs-lite"
 	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert/yaml"
 	cli "github.com/urfave/cli/v3"
@@ -247,15 +246,6 @@ const (
 	NVIDIAFS      = "nvidia-fs"
 	NVIDIAPEERMEM = "nvidia-peermem"
 )
-
-var hostNvidiaSMISearchPaths = []string{
-	"/usr/bin/nvidia-smi",
-	"/usr/sbin/nvidia-smi",
-	"/bin/nvidia-smi",
-	"/sbin/nvidia-smi",
-	wslNvidiaSMIPath,
-	"/opt/bin/nvidia-smi",
-}
 
 func main() {
 	c := cli.Command{}
@@ -738,36 +728,6 @@ func isDriverManagedByOperator(ctx context.Context) (bool, error) {
 	}
 
 	return false, nil
-}
-
-// resolveHostNvidiaSMI searches common nvidia-smi locations within the mounted
-// host root and returns the resolved path relative to the host root. A candidate
-// is only accepted if it resolves to a regular, non-empty, executable file, so
-// the privileged validator never execs a bogus binary from a non-standard path.
-func resolveHostNvidiaSMI(hostRootCtrPath string) (string, error) {
-	for _, nvidiaSMIPath := range hostNvidiaSMISearchPaths {
-		f, err := pathrs.OpenInRoot(hostRootCtrPath, nvidiaSMIPath)
-		if err != nil {
-			log.Debugf("failed to open '%s' on the host: %v", nvidiaSMIPath, err)
-			continue
-		}
-
-		fileInfo, err := f.Stat()
-		_ = f.Close()
-		if err != nil {
-			log.Debugf("failed to stat '%s' on the host: %v", nvidiaSMIPath, err)
-			continue
-		}
-
-		if !fileInfo.Mode().IsRegular() || fileInfo.Size() == 0 || fileInfo.Mode().Perm()&0o111 == 0 {
-			log.Debugf("skipping '%s' on the host: not a non-empty executable regular file", nvidiaSMIPath)
-			continue
-		}
-
-		return nvidiaSMIPath, nil
-	}
-
-	return "", fmt.Errorf("failed to find an executable 'nvidia-smi' on the host")
 }
 
 func validateHostDriver(silent bool) error {
